Refactor observation system (#339)
* update obs

* move state obs

* format

* make side detector better

* format

* remove use_lane_line_detector config

* format

* fix error

* format

* fix tests

* fix bug

* format
