+++ pgdrive/obs/top_down_obs.py
@@ -1,16 +1,19 @@
+"""
+We implement this top-down view render based on eleurent/Highway-Env
+See more information on its Github page: https://github.com/eleurent/highway-env
+"""
 import os
 import sys
 from typing import Tuple
 
 import gym
 import numpy as np
-
 from pgdrive.constants import Decoration, DEFAULT_AGENT, PG_EDITION
 from pgdrive.obs.observation_type import ObservationType
+from pgdrive.obs.top_down_obs_impl import WorldSurface, ObservationWindow, COLOR_BLACK, \
+    VehicleGraphics, LaneGraphics
 from pgdrive.scene_creator.vehicle.base_vehicle import BaseVehicle
 from pgdrive.utils import import_pygame
-from pgdrive.world.top_down_observation.top_down_obs_impl import WorldSurface, ObservationWindow, COLOR_BLACK, \
-    VehicleGraphics, LaneGraphics
 
 pygame = import_pygame()
