+++ pgdrive/tests/test_funtionality/test_expert_performance.py
@@ -37,7 +37,7 @@ def test_expert_with_traffic(use_render=False):
         dict(environment_num=1, start_seed=3, load_map_from_json=False, random_traffic=False, use_render=use_render),
         num_episode=3
     )
-    assert 315 < ep_reward < 335, ep_reward
+    assert 315 < ep_reward < 345, ep_reward
     assert success_rate == 1.0, success_rate
 
 
@@ -46,7 +46,7 @@ def test_expert_without_traffic():
         dict(environment_num=1, start_seed=0, traffic_density=0, load_map_from_json=False, random_traffic=False),
         num_episode=3
     )
-    assert 250 <= ep_reward <= 265, ep_reward
+    assert 250 <= ep_reward <= 285, ep_reward
     assert success_rate == 1.0, success_rate
