+++ pgdrive/tests/vis_env/vis_pgdrive_env.py
@@ -37,7 +37,6 @@ class TestEnv(PGDriveEnv):
                     "show_lidar": False,
                     "show_side_detector": True,
                     "show_lane_line_detector": True,
-                    "use_lane_line_detector": True,
                 }
             }
         )
